Some code refactoring and adding comments
