+++ pom.xml
@@ -12,9 +12,11 @@
         <maven.compiler.source>11</maven.compiler.source>
         <maven.compiler.target>11</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        
         <lombok.version>1.18.24</lombok.version>
         <guava.version>31.1-jre</guava.version>
         <commons-imaging.version>1.0-alpha2</commons-imaging.version>
+        <commons-io.version>2.7</commons-io.version>
     </properties>
 
     <dependencies>
@@ -31,7 +33,7 @@
         <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
-            <version>2.7</version>
+            <version>${commons-io.version}</version>
         </dependency>
         <dependency>
             <groupId>org.apache.commons</groupId>
