+++ src/main/java/com/hescha/parser/sff/model/v2/SffV2Item.java
@@ -14,11 +14,12 @@ public class SffV2Item {
     private int imageAxisY;
     private int linkedSpriteNumber;
     /**
-     *     1 invalid (no use)
-     *     2 RLE8
-     *     3 RLE5
-     *     4 LZ5
+     * 1 invalid (no use)
+     * 2 RLE8
+     * 3 RLE5
+     * 4 LZ5
      */
+    // TODO: create enum for this field
     private int compressionAlgorithm;
     private int biDepth;
     private int offset;
